integration: make Go tests available as a package
Any package can now have Go tests by using vmtest.GolangTest on itself.
