+++ pkg/vmtest/internal/json2test/testcollector.go
